+++ lib/serega/plugins/batch/lib/validations/check_batch_opt_loader.rb
@@ -22,45 +22,15 @@ class Serega
 
             raise SeregaError, must_be_callable unless loader.respond_to?(:call)
 
-            if loader.is_a?(Proc)
-              check_block(loader)
-            else
-              check_callable(loader)
-            end
+            SeregaValidations::Utils::CheckExtraKeywordArg.call(:loader, loader)
+            params_count = SeregaUtils::ParamsCount.call(loader, max_count: 3)
+            raise SeregaError, params_count_error if (params_count != 1) && (params_count != 2) && (params_count != 3)
           end
 
           private
 
-          def check_block(block)
-            return if valid_parameters?(block, accepted_count: 0..3)
-
-            raise SeregaError, block_parameters_error
-          end
-
-          def check_callable(callable)
-            return if valid_parameters?(callable.method(:call), accepted_count: 3..3)
-
-            raise SeregaError, callable_parameters_error
-          end
-
-          def valid_parameters?(data, accepted_count:)
-            params = data.parameters
-            accepted_count.include?(params.count) && valid_parameters_types?(params)
-          end
-
-          def valid_parameters_types?(params)
-            params.all? do |param|
-              type = param[0]
-              (type == :req) || (type == :opt)
-            end
-          end
-
-          def block_parameters_error
-            "Invalid :batch option :loader. When it is a Proc it can have maximum three regular parameters (keys, context, point)"
-          end
-
-          def callable_parameters_error
-            "Invalid :batch option :loader. When it is a callable object it must have three regular parameters (keys, context, point)"
+          def params_count_error
+            "Invalid :batch option :loader. It must accept 1, 2 or 3 parameters (keys, context, plan)"
           end
 
           def must_be_callable