+++ lib/serega/plugins/if/validations/check_opt_unless_value.rb
@@ -26,52 +26,20 @@ class Serega
 
           private
 
-          def check_type(value)
-            return if value.is_a?(Symbol)
-
-            raise SeregaError, must_be_callable unless value.respond_to?(:call)
-
-            if value.is_a?(Proc)
-              check_block(value)
-            else
-              check_callable(value)
-            end
-          end
-
           def check_usage_with_other_params(opts)
             raise SeregaError, "Option :unless_value can not be used together with option :serializer" if opts.key?(:serializer)
           end
 
-          def check_block(block)
-            return if valid_parameters?(block, accepted_count: 0..2)
+          def check_type(value)
+            return if value.is_a?(Symbol)
+            raise SeregaError, must_be_callable unless value.respond_to?(:call)
 
-            raise SeregaError, block_parameters_error
-          end
+            SeregaValidations::Utils::CheckExtraKeywordArg.call(:unless_value, value)
+            params_count = SeregaUtils::ParamsCount.call(value, max_count: 2)
 
-          def check_callable(callable)
-            return if valid_parameters?(callable.method(:call), accepted_count: 2..2)
-
-            raise SeregaError, callable_parameters_error
-          end
-
-          def valid_parameters?(data, accepted_count:)
-            params = data.parameters
-            accepted_count.include?(params.count) && valid_parameters_types?(params)
-          end
-
-          def valid_parameters_types?(params)
-            params.all? do |param|
-              type = param[0]
-              (type == :req) || (type == :opt)
+            if params_count > 2
+              raise SeregaError, "Option :unless_value value should have up to 2 parameters (value, context)"
             end
-          end
-
-          def block_parameters_error
-            "Invalid attribute option :unless_value. When it is a Proc it can have maximum two regular parameters (object, context)"
-          end
-
-          def callable_parameters_error
-            "Invalid attribute option :unless_value. When it is a callable object it must have two regular parameters (object, context)"
           end
 
           def must_be_callable