+++ spec/serega/plugins/batch/lib/validations/check_batch_opt_loader_spec.rb
@@ -3,12 +3,12 @@
 load_plugin_code :batch
 
 RSpec.describe Serega::SeregaPlugins::Batch::CheckBatchOptLoader do
-  let(:block_parameters_error) do
-    "Invalid :batch option :loader. When it is a Proc it can have maximum three regular parameters (keys, context, point)"
+  let(:params_count_error) do
+    "Invalid :batch option :loader. It must accept 1, 2 or 3 parameters (keys, context, plan)"
   end
 
-  let(:callable_parameters_error) do
-    "Invalid :batch option :loader. When it is a callable object it must have three regular parameters (keys, context, point)"
+  let(:param_type_error) do
+    "Option :loader value should not accept keyword argument `a:`"
   end
 
   let(:must_be_callable) do
@@ -24,56 +24,26 @@ RSpec.describe Serega::SeregaPlugins::Batch::CheckBatchOptLoader do
     expect { described_class.call(Object.new) }.to raise_error Serega::SeregaError, must_be_callable
   end
 
-  it "allows Proc with 0-3 regular args" do
-    expect { described_class.call(proc {}) }.not_to raise_error
-    expect { described_class.call(proc { |a| }) }.not_to raise_error
-    expect { described_class.call(proc { |a, b| }) }.not_to raise_error
-    expect { described_class.call(proc { |a, b, c| }) }.not_to raise_error
-    expect { described_class.call(proc { |a, b, c, d| }) }.to raise_error Serega::SeregaError, block_parameters_error
-    expect { described_class.call(proc { |a:| }) }.to raise_error Serega::SeregaError, block_parameters_error
-    expect { described_class.call(proc { |*a| }) }.to raise_error Serega::SeregaError, block_parameters_error
+  it "prohibits to add loader with required keyword args" do
+    value = proc { |a:| }
+    expect { described_class.call(value) }.to raise_error Serega::SeregaError, param_type_error
+  end
+
+  it "allows Proc with 1 to 3 args" do
+    value = proc {}
+    counter = Serega::SeregaUtils::ParamsCount
+    allow(counter).to receive(:call).and_return(0, 1, 2, 3, 4)
+
+    expect { described_class.call(value) }.to raise_error Serega::SeregaError, params_count_error
+    expect { described_class.call(value) }.not_to raise_error
+    expect { described_class.call(value) }.not_to raise_error
+    expect { described_class.call(value) }.not_to raise_error
+    expect { described_class.call(value) }.to raise_error Serega::SeregaError, params_count_error
+
+    expect(counter).to have_received(:call).with(value, max_count: 3).exactly(5).times
   end
 
   it "allows symbols" do
     expect { described_class.call(:foo) }.not_to raise_error
-  end
-
-  it "allows callable value with 3 regular args" do
-    callable0 = Class.new {
-      def self.call
-      end
-    }
-    callable1 = Class.new {
-      def self.call(a)
-      end
-    }
-    callable2 = Class.new {
-      def self.call(a, b)
-      end
-    }
-    callable3 = Class.new {
-      def self.call(a, b, c)
-      end
-    }
-    callable4 = Class.new {
-      def self.call(a, b, c, d)
-      end
-    }
-    callable5 = Class.new {
-      def self.call(a:, b:, c:)
-      end
-    }
-    callable6 = Class.new {
-      def self.call(*a)
-      end
-    }
-
-    expect { described_class.call(callable0) }.to raise_error Serega::SeregaError, callable_parameters_error
-    expect { described_class.call(callable1) }.to raise_error Serega::SeregaError, callable_parameters_error
-    expect { described_class.call(callable2) }.to raise_error Serega::SeregaError, callable_parameters_error
-    expect { described_class.call(callable3) }.not_to raise_error
-    expect { described_class.call(callable4) }.to raise_error Serega::SeregaError, callable_parameters_error
-    expect { described_class.call(callable5) }.to raise_error Serega::SeregaError, callable_parameters_error
-    expect { described_class.call(callable6) }.to raise_error Serega::SeregaError, callable_parameters_error
   end
 end